style: rename the package
