+++ src/main/kotlin/bass/controller/member/MemberController.kt
@@ -2,7 +2,7 @@ package bass.controller.member
 
 import bass.controller.member.usecase.AuthUseCase
 import bass.controller.member.usecase.CrudMemberUseCase
-import bass.dto.MemberRegisterDTO
+import bass.dto.member.MemberRegisterDTO
 import bass.dto.TokenRequestDTO
 import bass.dto.TokenResponseDTO
 import bass.infrastructure.AuthorizationExtractor
