+++ src/main/kotlin/bass/controller/order/usecase/OrderCreationUseCase.kt
@@ -1,6 +1,6 @@
 package bass.controller.order.usecase
 
-import bass.dto.MemberLoginDTO
+import bass.dto.member.MemberLoginDTO
 import bass.dto.OrderDTO
 import bass.model.PaymentRequest
