+++ src/main/kotlin/bass/dto/member/MemberLoginDTO.kt
@@ -1,4 +1,4 @@
-package bass.dto
+package bass.dto.member
 
 data class MemberLoginDTO(
     var id: Long,
