+++ src/main/kotlin/bass/services/order/OrderServiceImpl.kt
@@ -1,7 +1,7 @@
 package bass.services.order
 
 import bass.controller.order.usecase.OrderCreationUseCase
-import bass.dto.MemberLoginDTO
+import bass.dto.member.MemberLoginDTO
 import bass.dto.OrderDTO
 import bass.dto.PaymentDTO
 import bass.entities.CartItemEntity
